Fix setTagManager method to tag onto setPersons
